app/requirements.txt
APScheduler==3.10.4
DateTime==5.5
greenlet==3.0.3
memory-profiler==0.61.0
numpy==2.0.0
pandas==2.2.2
psutil==6.0.0
psycopg2==2.9.9
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
zope.interface==6.4.post2
